Move actions code to separate file, add error: False to successful return value
